feat(cv): vue mobile indépendante — /short affiche le CV complet sur mobile (ADR-0006)
Implémente clean-B (job-app ADR-0006) par RÉUTILISATION : sur mobile, /[lang]/short
n'affiche plus le CV court mais la vue COMPLÈTE (identique à /[lang]), sans redirect
ni réécriture d'URL. Le choix court/complet ne se fait qu'à l'impression (modale).

- Retire MobileShortToFullRedirect (+ son spec) : plus de router.replace, l'URL
  /short reste /short → l'aller-retour de partage (desktop→mobile→desktop) est préservé.
- /short rendu en 2 branches CSS-exclusives :
  · MOBILE (md:hidden print:hidden print-preview:hidden) → <OfferTailoredShell> (complet)
  · DESKTOP+IMPRESSION (hidden md:block print:block print-preview:block) → CV court A4 (inchangé)
- Enveloppe A4 (cv-print-preview + cv-short-page + zoom) déplacée du layout vers la
  branche desktop : la vue mobile ne doit pas hériter du cadrage A4 du court.
- WYSIWYG intact : l'arbre A4 (aperçu ?print + PDF) n'est pas modifié ; la vue mobile
  est print:hidden (jamais imprimée).
- Effets globaux (print-preview, ATS, autoprint) montés une seule fois au niveau page ;
  OfferTailoredShell reçoit renderEffects={false} en sous-vue pour éviter le double-montage.
- lib/full-cv-shell-props.ts : source unique des props du CV long, partagée / et /short.

Reste (suivi) : adapter ~9 specs e2e couplées + scoping des id (collision #domains/
#studies/#projects entre les 2 arbres) + purge des max-md: morts de l'arbre A4.
